pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "selfology-bot"
version = "0.1.0"
description = "Selfology AI Psychology Coach - FastAPI + Telegram Bot"
authors = [{name = "ksnk"}]
dependencies = [
    "fastapi>=0.104.0",
    "aiogram>=3.2.0",
    "anthropic>=0.7.0",
    "openai>=1.6.0",
    "qdrant-client>=1.7.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]